rename: entity -> domain
