+++ src/main/java/com/nhnacademy/bookingservice/domain/Booking.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.bookingservice.entity;
+package com.nhnacademy.bookingservice.domain;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
