+++ src/main/java/com/nhnacademy/bookingservice/repository/BookingChangeRepository.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.bookingservice.repository;
 
-import com.nhnacademy.bookingservice.entity.BookingChange;
+import com.nhnacademy.bookingservice.domain.BookingChange;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookingChangeRepository extends JpaRepository<BookingChange, Long> {
