+++ src/main/java/com/nhnacademy/bookingservice/repository/impl/CustomBookingRepositoryImpl.java
@@ -4,10 +4,10 @@ import com.nhnacademy.bookingservice.dto.BookingResponse;
 import com.nhnacademy.bookingservice.dto.DailyBookingResponse;
 import com.nhnacademy.bookingservice.dto.QBookingResponse;
 import com.nhnacademy.bookingservice.dto.QDailyBookingResponse;
-import com.nhnacademy.bookingservice.entity.Booking;
-import com.nhnacademy.bookingservice.entity.BookingChangeType;
-import com.nhnacademy.bookingservice.entity.QBooking;
-import com.nhnacademy.bookingservice.entity.QBookingChange;
+import com.nhnacademy.bookingservice.domain.Booking;
+import com.nhnacademy.bookingservice.domain.BookingChangeType;
+import com.nhnacademy.bookingservice.domain.QBooking;
+import com.nhnacademy.bookingservice.domain.QBookingChange;
 import com.nhnacademy.bookingservice.repository.CustomBookingRepository;
 import com.nhnacademy.bookingservice.repository.util.QueryDslUtil;
 import com.querydsl.core.BooleanBuilder;
