+++ src/test/java/com/nhnacademy/bookingservice/service/impl/BookingServiceImplTest.java
@@ -10,9 +10,9 @@ import com.nhnacademy.bookingservice.common.exception.booking.*;
 import com.nhnacademy.bookingservice.common.exception.NotFoundException;
 import com.nhnacademy.bookingservice.common.exception.meeting.MeetingRoomCapacityExceededException;
 import com.nhnacademy.bookingservice.dto.*;
-import com.nhnacademy.bookingservice.entity.Booking;
-import com.nhnacademy.bookingservice.entity.BookingChange;
-import com.nhnacademy.bookingservice.entity.BookingChangeType;
+import com.nhnacademy.bookingservice.domain.Booking;
+import com.nhnacademy.bookingservice.domain.BookingChange;
+import com.nhnacademy.bookingservice.domain.BookingChangeType;
 import com.nhnacademy.bookingservice.repository.BookingChangeRepository;
 import com.nhnacademy.bookingservice.repository.BookingRepository;
 import org.junit.jupiter.api.*;
